F-11 B1+B2: break the modeling import cycle, hoist base.py imports
Re-point the four _evaluation modules to import SklearnBridge from the
leaf engines.sklearn_bridge instead of modeling.sklearn_wrapper, closing
the only real cycle (base -> _evaluation -> sklearn_wrapper -> base).
Hoist the two deferred import sites in base.py (perform_cross_validation
and the three evaluate_* dispatchers) to module level so failures surface
at import time. Fresh-process import probes and full core suite green.
